+++ src/main/java/wooteco/subway/service/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.service.dto;
 
 import wooteco.subway.domain.Line;
