+++ src/main/java/wooteco/subway/service/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.service.dto;
 
 import javax.validation.constraints.NotBlank;
