+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.StationResponse;
+import wooteco.subway.service.dto.StationResponse;
 
 @DisplayName("지하철역 관련 인수테스트")
 public class StationAcceptanceTest extends AcceptanceTest {
