+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -15,7 +15,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
-import wooteco.subway.dto.LineRequest;
+import wooteco.subway.service.dto.LineRequest;
 
 import java.util.List;
 import java.util.Optional;
